authentication ok

+++ tools/copy-files.mjs
@@ -8,17 +8,18 @@ fs.cpSync(ssr, worker, { recursive: true });
 fs.cpSync(join(cloudflare, 'index.html'), join(cloudflare, '404.html'));
 fs.cpSync(join(cloudflare, 'index.html'), join(worker, '404.html'));
 
-const data = fs.readFileSync(join(ssr, '../prerendered-routes.json'));
-if (!fs.existsSync(ssg)) {
-  fs.mkdirSync(ssg);
+if (fs.existsSync(join(ssr, '../prerendered-routes.json'))) {
+  const data = fs.readFileSync();
+  if (!fs.existsSync(ssg)) {
+    fs.mkdirSync(ssg);
+  }
+
+  const routes = JSON.parse(data).routes;
+  routes
+    .map((r) => (r.indexOf('/', 2) > 0 ? undefined : r.substring(1)))
+    .filter((r) => r !== undefined && r !== '/' && r !== '')
+    .forEach((path) => {
+      fs.renameSync(join(cloudflare, path), join(ssg, path));
+    });
 }
-
-const routes = JSON.parse(data).routes;
-routes
-  .map((r) => (r.indexOf('/', 2) > 0 ? undefined : r.substring(1)))
-  .filter((r) => r !== undefined && r !== '/' && r !== '')
-  .forEach((path) => {
-    fs.renameSync(join(cloudflare, path), join(ssg, path));
-  });
-
 fs.renameSync(join(worker, 'server.mjs'), join(worker, 'index.js'));
